Refactor env settings and add secrets rotation automation
